# backend/swagger.yaml
...
openapi: 3.0.3
info:
  title: CIT Event Hub API
  version: 1.0.0
  description: >-
    Comprehensive API documentation for the CIT Event Hub backend. Includes endpoints for authentication, user management, events, teams, registrations, notifications, payments, results, and announcements.
servers:
  - url: http://localhost:3000/api-docs
    description: Local development server
paths:
...
  /users:
    get:
      summary: List all users (admin only)
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
              example:
                - id: 1
                  email: admin@example.com
                  role: admin
                - id: 2
                  email: student@example.com
                  role: student
...
  /teams:
    post:
      summary: Create a team (student, organizer, admin)
      tags:
        - Teams
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
            example:
              event_id: 5
              name: "Alpha Squad"
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
...
  /notifications:
    post:
      summary: Create a notification (admin/system)
      tags:
        - Notifications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationCreate'
            example:
              user_id: 12
              type: announcement
              title: "New Announcement"
              body: "Check out the latest updates."
      responses:
        '201':
          description: Notification created
...
  /announcements:
    get:
      summary: List announcements (public)
      tags:
        - Announcements
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Announcement'
    post:
      summary: Create announcement (admin/organizer)
      tags:
        - Announcements
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementCreate'
            example:
              title: "Campus Event Tomorrow"
              body: "Don't miss the guest lecture at 10 AM."
      responses:
        '201':
          description: Announcement created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        created_at:
          type: string
          format: date-time
...
    Team:
      type: object
      properties:
        id:
          type: integer
        event_id:
          type: integer
        name:
          type: string
        captain_student_id:
          type: integer
    TeamCreate:
      type: object
      required:
        - event_id
        - name
      properties:
        event_id:
          type: integer
        name:
          type: string
        captain_student_id:
          type: integer
...
    NotificationCreate:
      type: object
      required:
        - user_id
        - type
        - title
      properties:
        user_id:
          type: integer
        type:
          type: string
          enum: [team_invite, event_reminder, announcement, payment_success, result_published]
        title:
          type: string
        body:
          type: string
        url:
          type: string
    Announcement:
      type: object
      properties:
        id:
          type: integer
        creator_user_id:
          type: integer
        title:
          type: string
        body:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AnnouncementCreate:
      type: object
      required:
        - title
        - body
      properties:
        title:
          type: string
        body:
          type: string
...